requirements.txt
easyocr
pandas
python-docx
numpy
openpyxl
Pillow
PyMuPDF
requests
torch
torchvision
xlrd
